Bug #1321803 soc: sprd: iommu: fix incorrect variable type for map_cnt
[root cause  ] iommu map counter overflow
[changes     ] use u32 for map_cnt
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:sharkl5pro
[test case]:vdsp stress
[reviewers   ] Wenan Hu
[change_type ] feature_bugfix
[tag_product ] specific

Fix incorrect variable type for map_cnt.

+++ drivers/soc/sprd/iommu/drv/cll/sprd_iommuex_cll.h
@@ -49,7 +49,7 @@ struct sprd_iommuex_priv {
 	unsigned int pagt_ddr_size;
 	unsigned long pagt_base_phy_ddr;
 
-	u8 map_cnt;
+	u32 map_cnt;
 	enum IOMMU_ID iommu_id;
 };
 #endif  /* _SPRD_IOMMU_CLL_H_ */
